--- reversi-server/services/src/main/java/io/deeplay/camp/bot/BotStrategy.java
@@ -1,21 +1,0 @@
-package io.deeplay.camp.bot;
-
-import io.deeplay.camp.entity.Tile;
-import io.deeplay.camp.board.BoardService;
-import org.jetbrains.annotations.NotNull;
-
-import java.util.List;
-
-public abstract class BotStrategy {
-    public final int id;
-    public final String name;
-
-    protected BotStrategy(int id, String name) {
-        this.id = id;
-        this.name = name;
-    }
-
-    public abstract Tile getMove(int currentPlayerId, @NotNull BoardService boardLogic);
-
-    public abstract List<Tile> getAllValidMoves(int currentPlayerId, @NotNull BoardService boardLogic);
-}
